+++ docs/CHANGELOG.md
@@ -1,12 +1,43 @@
 # Change Log
-All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+All notable changes to the "vscode-rainbow-fart" extension will be documented in this file.
 
-<!--
-## [Unreleased]
--->
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [1.0.0] - 2020-06-14
+## [1.2.0] - 2020-06-20
 
-Initial Release
+### Added
+
+- Enable/Disable voice packages.
+- English document.
+
+### Fixed
+
+- Duplicated key in the voice package details modal.
+- Fixed width of voice package details modal
+- Issue: [#8](https://github.com/SaekiRaku/vscode-rainbow-fart/issues/8)
+
+## [1.1.0] - 2020-06-14
+
+### Added
+
+- Support import voice packages
+- New built-in voices for Chinese
+- I18N supports for English
+- Add document
+
+### Changed
+
+- New user interface
+
+## [1.0.2] - 2020-06-01
+
+- Lower the required version of VSCode.
+
+## [1.0.1] - 2020-05-30
+
+- UI Optimization
+
+## [1.0.0] - 2020-05-29
+
+- Initial release